Add files via upload
The previous version was intended primarily for people to tinker with the code, and the only on-screen controls were play/pause and randomize alpha, beta and gamma. The latest version is fully interactive. No editing is required but hopefully the now quite extensive code is still pretty accessible if people want to modify it.

I've only tested it on Chrome (desktop and mobile) and Internet Explorer (doesn't work).

Let me know what you think. Cheers.